PM: HACK: fix watchdog timer in suspend
The watchdog timer cannot be stopped in N9 once it has
been started. Hence the normal watcdog disable does not
work in suspend.

Created a hacky workaround to start a always-on timer
to wake up the system just before the watchdog would reset
the device.

Patch based on a debug patch for suspend wake up timers
from Santosh Shilimkar.

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jollamobile.com>