Vibra: Convert N9 vibra_spi to ff_memless
Converted the vibra_spi to use ff_memless driver instead of
directly exposing only periodic vibra driver for custom wave forms.

The N9 vibra is a special linear vibrator, that needs 150Hz sine
signal to create maximum force feedback effect. The signal is
generated as a PWM from SPI bus data out that is amplified then
to for the vibra.

Now the converted driver creates the 32 bit PWM sample words based
on rumble magnitudes from ff-memless and a pre-gerenated sine table.
The amplitude of the sample is measured in number of '1' bits in
the word:
- 0xffff ffff == 1
- 0x0000 ffff == 0.5
- 0x0000 0000 == 0
- etc.

For the 150Hz effective sine signal, 156 32bit samples are created, which
are then written to the SPI bus at 750kHz. Each call to the playback
function will generate new set of samples that replace the old ones.
Playback continues until effect with 0 rumble magnitudes is received.

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jollamobile.com>